requirements.txt
Flask>=0.10
pytz>=2015
awesome-slugify>=1.6.5
